@@ -22,7 +22,7 @@ import 'utils/location_utils.dart';
 class LocationPicker extends StatefulWidget {
   LocationPicker(
     this.apiKey, {
-    Key? key,
+    Key key,
     this.initialCenter,
     this.initialZoom,
     this.requiredGPS,
@@ -40,50 +40,45 @@ class LocationPicker extends StatefulWidget {
     this.countries,
     this.language,
     this.desiredAccuracy,
-    this.buttonColor,
-    this.buttonIcon,
-    this.backButton,
   });
 
   final String apiKey;
 
-  final LatLng? initialCenter;
-  final double? initialZoom;
-  final List<String>? countries;
+  final LatLng initialCenter;
+  final double initialZoom;
+  final List<String> countries;
 
-  final bool? requiredGPS;
-  final bool? myLocationButtonEnabled;
-  final bool? layersButtonEnabled;
-  final bool? automaticallyAnimateToCurrentLocation;
+  final bool requiredGPS;
+  final bool myLocationButtonEnabled;
+  final bool layersButtonEnabled;
+  final bool automaticallyAnimateToCurrentLocation;
 
-  final String? mapStylePath;
+  final String mapStylePath;
 
-  final Color? appBarColor;
-  final BoxDecoration? searchBarBoxDecoration;
-  final String? hintText;
-  final Widget? resultCardConfirmIcon;
-  final Alignment? resultCardAlignment;
-  final Decoration? resultCardDecoration;
-  final EdgeInsets? resultCardPadding;
+  final Color appBarColor;
+  final BoxDecoration searchBarBoxDecoration;
+  final String hintText;
+  final Widget resultCardConfirmIcon;
+  final Alignment resultCardAlignment;
+  final Decoration resultCardDecoration;
+  final EdgeInsets resultCardPadding;
 
-  final String? language;
+  final String language;
 
-  final LocationAccuracy? desiredAccuracy;
-  final Color? buttonColor;
-  final Widget? buttonIcon;
-  final Widget? backButton;
+  final LocationAccuracy desiredAccuracy;
+
   @override
   LocationPickerState createState() => LocationPickerState();
 }
 
 class LocationPickerState extends State<LocationPicker> {
   /// Result returned after user completes selection
-  LocationResult? locationResult;
+  LocationResult locationResult;
 
   /// Overlay to display autocomplete suggestions
-  OverlayEntry? overlayEntry;
+  OverlayEntry overlayEntry;
 
-  List<NearbyPlace> nearbyPlaces = [];
+  List<NearbyPlace> nearbyPlaces = List();
 
   /// Session token required for autocomplete API call
   String sessionToken = Uuid().generateV4();
@@ -99,7 +94,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// Hides the autocomplete overlay
   void clearOverlay() {
     if (overlayEntry != null) {
-      overlayEntry!.remove();
+      overlayEntry.remove();
       overlayEntry = null;
     }
   }
@@ -109,21 +104,22 @@ class LocationPickerState extends State<LocationPicker> {
   /// is hidden so as to give more room and better experience for the
   /// autocomplete list overlay.
   void searchPlace(String place) {
+    if (context == null) return;
+
     clearOverlay();
 
     setState(() => hasSearchTerm = place.length > 0);
 
     if (place.length < 1) return;
 
-    final RenderBox renderBox = context.findRenderObject() as RenderBox;
+    final RenderBox renderBox = context.findRenderObject();
     Size size = renderBox.size;
 
-    final RenderBox? appBarBox =
-        appBarKey.currentContext!.findRenderObject() as RenderBox?;
+    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
-        top: appBarBox!.size.height,
+        top: appBarBox.size.height,
         width: size.width,
         child: Material(
           elevation: 1,
@@ -150,7 +146,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context)!.insert(overlayEntry!);
+    Overlay.of(context).insert(overlayEntry);
 
     autoCompleteSearch(place);
   }
@@ -163,7 +159,7 @@ class LocationPickerState extends State<LocationPicker> {
 
     // Currently, you can use components to filter by up to 5 countries. from https://developers.google.com/places/web-service/autocomplete
     String regionParam = countries?.isNotEmpty == true
-        ? "&components=country:${countries!.sublist(0, min(countries.length, 5)).join('|country:')}"
+        ? "&components=country:${countries.sublist(0, min(countries.length, 5)).join('|country:')}"
         : "";
 
     var endpoint =
@@ -173,13 +169,12 @@ class LocationPickerState extends State<LocationPicker> {
             "language=${widget.language}";
 
     if (locationResult != null) {
-      endpoint += "&location=${locationResult!.latLng!.latitude}," +
-          "${locationResult!.latLng!.longitude}";
+      endpoint += "&location=${locationResult.latLng.latitude}," +
+          "${locationResult.latLng.longitude}";
     }
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint),
-            headers: headers as Map<String, String>?))
+        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> data = jsonDecode(response.body);
@@ -219,7 +214,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// To navigate to the selected place from the autocomplete list to the map,
   /// the lat,lng is required. This method fetches the lat,lng of the place and
   /// proceeds to moving the map to that location.
-  void decodeAndSelectPlace(String? placeId) {
+  void decodeAndSelectPlace(String placeId) {
     clearOverlay();
 
     final endpoint =
@@ -228,8 +223,7 @@ class LocationPickerState extends State<LocationPicker> {
             '&language=${widget.language}';
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint),
-            headers: headers as Map<String, String>?))
+        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> location =
@@ -246,18 +240,17 @@ class LocationPickerState extends State<LocationPicker> {
 
   /// Display autocomplete suggestions with the overlay.
   void displayAutoCompleteSuggestions(List<RichSuggestion> suggestions) {
-    final RenderBox renderBox = context.findRenderObject() as RenderBox;
+    final RenderBox renderBox = context.findRenderObject();
     Size size = renderBox.size;
 
-    final RenderBox? appBarBox =
-        appBarKey.currentContext!.findRenderObject() as RenderBox?;
+    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
 
     clearOverlay();
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
         width: size.width,
-        top: appBarBox!.size.height,
+        top: appBarBox.size.height,
         child: Material(
           elevation: 1,
           child: Column(
@@ -267,7 +260,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context)!.insert(overlayEntry!);
+    Overlay.of(context).insert(overlayEntry);
   }
 
   /// Utility function to get clean readable name of a location. First checks
@@ -299,8 +292,7 @@ class LocationPickerState extends State<LocationPicker> {
               "location=${latLng.latitude},${latLng.longitude}&radius=150" +
               "&language=${widget.language}";
 
-      return http.get(Uri.parse(endpoint),
-          headers: headers as Map<String, String>?);
+      return http.get(Uri.parse(endpoint), headers: headers);
     }).then((response) {
       if (response.statusCode == 200) {
         nearbyPlaces.clear();
@@ -338,15 +330,14 @@ class LocationPickerState extends State<LocationPicker> {
             "&language=${widget.language}";
 
     final response = await http.get(Uri.parse(endpoint),
-        headers: await (LocationUtils.getAppHeaders()
-            as FutureOr<Map<String, String>?>));
+        headers: await LocationUtils.getAppHeaders());
 
     if (response.statusCode == 200) {
       Map<String, dynamic> responseJson = jsonDecode(response.body);
 
-      String? road;
+      String road;
 
-      String? placeId = responseJson['results'][0]['place_id'];
+      String placeId = responseJson['results'][0]['place_id'];
 
       if (responseJson['status'] == 'REQUEST_DENIED') {
         road = 'REQUEST DENIED = please see log for more details';
@@ -361,9 +352,9 @@ class LocationPickerState extends State<LocationPicker> {
 
       setState(() {
         locationResult = LocationResult();
-        locationResult!.address = road;
-        locationResult!.latLng = latLng;
-        locationResult!.placeId = placeId;
+        locationResult.address = road;
+        locationResult.latLng = latLng;
+        locationResult.placeId = placeId;
       });
     }
   }
@@ -371,7 +362,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// Moves the camera to the provided location and updates other UI features to
   /// match the location.
   void moveToLocation(LatLng latLng) {
-    mapKey.currentState!.mapController.future.then((controller) {
+    mapKey.currentState.mapController.future.then((controller) {
       controller.animateCamera(
         CameraUpdate.newCameraPosition(
           CameraPosition(
@@ -406,7 +397,6 @@ class LocationPickerState extends State<LocationPicker> {
             iconTheme: Theme.of(context).iconTheme,
             elevation: 0,
             backgroundColor: widget.appBarColor,
-            leading: widget.backButton,
             key: appBarKey,
             title: SearchInput(
               (input) => searchPlace(input),
@@ -452,24 +442,24 @@ class LocationPickerState extends State<LocationPicker> {
 /// set [automaticallyAnimateToCurrentLocation] to false.
 ///
 ///
-Future<LocationResult?> showLocationPicker(
+Future<LocationResult> showLocationPicker(
   BuildContext context,
   String apiKey, {
   LatLng initialCenter = const LatLng(45.521563, -122.677433),
   double initialZoom = 16,
   bool requiredGPS = false,
-  List<String>? countries,
+  List<String> countries,
   bool myLocationButtonEnabled = false,
   bool layersButtonEnabled = false,
   bool automaticallyAnimateToCurrentLocation = true,
-  String? mapStylePath,
+  String mapStylePath,
   Color appBarColor = Colors.transparent,
-  BoxDecoration? searchBarBoxDecoration,
-  String? hintText,
-  Widget? resultCardConfirmIcon,
-  AlignmentGeometry? resultCardAlignment,
-  EdgeInsetsGeometry? resultCardPadding,
-  Decoration? resultCardDecoration,
+  BoxDecoration searchBarBoxDecoration,
+  String hintText,
+  Widget resultCardConfirmIcon,
+  AlignmentGeometry resultCardAlignment,
+  EdgeInsetsGeometry resultCardPadding,
+  Decoration resultCardDecoration,
   String language = 'en',
   LocationAccuracy desiredAccuracy = LocationAccuracy.best,
 }) async {
@@ -491,8 +481,8 @@ Future<LocationResult?> showLocationPicker(
           hintText: hintText,
           searchBarBoxDecoration: searchBarBoxDecoration,
           resultCardConfirmIcon: resultCardConfirmIcon,
-          resultCardAlignment: resultCardAlignment as Alignment?,
-          resultCardPadding: resultCardPadding as EdgeInsets?,
+          resultCardAlignment: resultCardAlignment,
+          resultCardPadding: resultCardPadding,
           resultCardDecoration: resultCardDecoration,
           countries: countries,
           language: language,
